tidy to align with coding standards
