--- src/main/java/Duke/task/Event.java
@@ -1,27 +1,0 @@
-package Duke.task;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-public class Event extends Task {
-    Date by;
-
-    public Event(String description, Date by) {
-        super(description);
-        this.by = by;
-    }
-
-    @Override
-    public String formatToWrite() {
-        if (this.done) {
-            return String.format("E | %d | %s | %s", 1, this.description, new SimpleDateFormat("dd/MM/yyyy HHmm").format(this.by));
-        } else {
-            return String.format("E | %d | %s | %s", 0, this.description, new SimpleDateFormat("dd/MM/yyyy HHmm").format(this.by));
-        }
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[E]%s (at: %s)", super.toString(), new SimpleDateFormat("dd/MM/yyyy HHmm").format(this.by));
-    }
-}
