+++ src/main/java/duke/exception/DukeException.java
@@ -1,4 +1,4 @@
-package Duke.exception;
+package duke.exception;
 
 public class DukeException extends Exception {
     String errorMsg;
