+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ Hello from
 |____/ \__,_|_|\_\___|
 
 	____________________________________________________________
-	 Hello! I'm Duke
+	 Hello! I'm duke
 	 What can I do for you?
 	____________________________________________________________
 	____________________________________________________________
